backend/requirements.txt
# Core
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1

# Data Processing
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
scipy==1.11.4
Pillow==10.3.0

# ML/DL
torch==2.0.1+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.2+cu118 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
transformers==4.35.2
sentence-transformers==2.2.2
accelerate==0.25.0

# Vector Store
faiss-cpu==1.7.4

# Document Processing
pymupdf==1.23.8

# Hugging Face
huggingface-hub==0.20.3

# Utils
tqdm==4.66.1

# Google Generative AI
google-generativeai==0.3.2
